Create Theorem-Prover.py
This is version 1 of Theorem Prover for Propositional Logic written in Python. It doesn't take a formula input from the user as of the moment but rather it gives the user options to choose from the given formulas.

I will be updating this code in the future such that it can get a formula input from the user in cnf form and then prove the given input whether it is Tautology, Contingency or Contradiction.